fixed sonarQube  errors and refactored code
