+++ .github/workflows/firebase-hosting-merge.yml
@@ -8,6 +8,8 @@ name: Deploy to Firebase Hosting on merge
             - main
 env:
     REACT_APP_DRIBBLE_ACCESS_TOKEN: ${{secrets.REACT_APP_DRIBBLE_ACCESS_TOKEN}}
+    REACT_APP_EMAIL_SERVICE_ID: ${{secrets.REACT_APP_EMAIL_SERVICE_ID}}
+    REACT_APP_EMAIL_TEMPLATE_ID: ${{secrets.REACT_APP_EMAIL_TEMPLATE_ID}}
 jobs:
     build_and_deploy:
         runs-on: ubuntu-latest
