+++ src/components/NavBar/DesktopNavBar.jsx
@@ -15,45 +15,42 @@ const useStyles = makeStyles((theme) => ({
 const DesktopNavBar = () => {
     const classes = useStyles();
     const activeStyle = {
-        // textDecoration: "underline solid #FE4747 2px",
         borderBottom: "2px solid #FE4747 ",
         borderRadius: "2px",
     };
     return (
-        <>
-            <AppBar position="static" color="transparent" elevation={0} className={classes.showDesktop}>
-                <Toolbar>
-                    <Typography style={{ flexGrow: "1" }}>
-                        <Link
-                            to="/
+        <AppBar position="static" color="transparent" elevation={0} className={classes.showDesktop}>
+            <Toolbar>
+                <Typography style={{ flexGrow: "1" }}>
+                    <Link
+                        to="/
                         ">
-                            <LogoComp />
-                        </Link>
-                    </Typography>
-                    <Button
-                        color="inherit"
-                        component={NavLink}
-                        to="/"
-                        style={({ isActive }) => (isActive ? activeStyle : undefined)}>
-                        Home
-                    </Button>
-                    <Button
-                        color="inherit"
-                        component={NavLink}
-                        to="/aboutme"
-                        style={({ isActive }) => (isActive ? activeStyle : undefined)}>
-                        About me
-                    </Button>
-                    <Button
-                        color="inherit"
-                        component={NavLink}
-                        to="/projects"
-                        style={({ isActive }) => (isActive ? activeStyle : undefined)}>
-                        Projects
-                    </Button>
-                </Toolbar>
-            </AppBar>
-        </>
+                        <LogoComp />
+                    </Link>
+                </Typography>
+                <Button
+                    color="inherit"
+                    component={NavLink}
+                    to="/"
+                    style={({ isActive }) => (isActive ? activeStyle : undefined)}>
+                    Home
+                </Button>
+                <Button
+                    color="inherit"
+                    component={NavLink}
+                    to="/aboutme"
+                    style={({ isActive }) => (isActive ? activeStyle : undefined)}>
+                    About me
+                </Button>
+                <Button
+                    color="inherit"
+                    component={NavLink}
+                    to="/projects"
+                    style={({ isActive }) => (isActive ? activeStyle : undefined)}>
+                    Projects
+                </Button>
+            </Toolbar>
+        </AppBar>
     );
 };
