+++ src/components/Pages/Home/Home.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { Box, Button, makeStyles } from "@material-ui/core";
 
-import "../../components/Pages/css/home.css";
+import "../../../css/home.css";
 import { faBehance, faDribbble, faGithub } from "@fortawesome/free-brands-svg-icons";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 
@@ -20,6 +20,7 @@ const useStyles = makeStyles((theme) => ({
         },
     },
 }));
+
 const Home = () => {
     const classes = useStyles();
     return (
@@ -29,10 +30,9 @@ const Home = () => {
                     <h3 className="title">Tapesh patel </h3>
                     <h1 className="header-1 slide-in-left "> Front-End developer</h1>
                     <h1 className="header-2 slide-in-right"> UI-UX Designer</h1>
-                    <p className="subtitile"> Why should you hire me for your next job ?</p>
+                    <p className="subtitle"> Why should you hire me for your next job ?</p>
                 </Box>
                 <Button className="more-btn" variant="contained" href="/aboutme">
-                    {" "}
                     Find more
                 </Button>
                 <Box className="home-icons">
@@ -68,9 +68,8 @@ const Home = () => {
                     <h3 className="m-title">Tapesh patel </h3>
                     <h1 className="m-header-1 slide-in-left "> Front-End developer</h1>
                     <h1 className="m-header-2 slide-in-right"> UI-UX Designer</h1>
-                    <p className="m-subtitile"> Why should you hire me for your next job ?</p>
+                    <p className="m-subtitle"> Why should you hire me for your next job ?</p>
                     <Button className="m-more-btn" href="/aboutme">
-                        {" "}
                         Find more
                     </Button>
                     <Box className="m-home-icons">
