+++ src/css/home.css
@@ -13,7 +13,6 @@
     text-transform: uppercase;
     font-size: 4em;
 
-    /* Create the gradient. */
     background-image: linear-gradient(
         45deg,
         #000000 10%,
@@ -29,7 +28,6 @@
         #a7489b 83.333%
     );
 
-    /* Set the background size and repeat properties. */
     background-size: 100%;
     background-repeat: repeat;
     -webkit-background-clip: text;
@@ -49,7 +47,6 @@
     text-transform: uppercase;
     font-size: 4em;
 
-    /* Create the gradient. */
     background-image: linear-gradient(
         45deg,
         #000000 10%,
@@ -65,7 +62,6 @@
         #a7489b 83.333%
     );
 
-    /* Set the background size and repeat properties. */
     background-size: 100%;
     background-repeat: repeat;
     -webkit-background-clip: text;
@@ -104,7 +100,7 @@
     height: 25px;
 }
 
-.subtitile {
+.subtitle {
     position: absolute;
     top: 65%;
     left: 15%;
@@ -142,8 +138,6 @@
 /* ----------------------for mobile  ------------------------*/
 
 .m-home-main {
-    /* background-color: aqua;
-    border: 1px solid green; */
     width: 70%;
     margin: auto;
     height: fit-content;
@@ -163,7 +157,6 @@
     text-transform: uppercase;
     font-size: 2.7rem;
 
-    /* Create the gradient. */
     background-image: linear-gradient(
         45deg,
         #000000 10.5%,
@@ -179,7 +172,6 @@
         #a7489b 83.333%
     );
 
-    /* Set the background size and repeat properties. */
     background-size: 100%;
     background-repeat: repeat;
     -webkit-background-clip: text;
@@ -211,7 +203,6 @@
     text-transform: uppercase;
     font-size: 2.7rem;
 
-    /* Create the gradient. */
     background-image: linear-gradient(
         45deg,
         #000000 10.5%,
@@ -227,7 +218,6 @@
         #a7489b 83.333%
     );
 
-    /* Set the background size and repeat properties. */
     background-size: 100%;
     background-repeat: repeat;
     -webkit-background-clip: text;
@@ -238,7 +228,7 @@
     animation: rainbow-text-simple-animation-rev 7s ease alternate-reverse infinite;
 }
 
-.m-subtitile {
+.m-subtitle {
     font-size: 1.2rem;
 }
